Fixed build error: Changed custom enum to int and passed the event.

+++ SteamVRLib_Valve.VR/SteamVR/Input/SteamVR_Behaviour_Pose.cs
@@ -170,7 +170,7 @@ namespace Valve.VR
 
                     if (broadcastDeviceChanges)
                     {
-                        this.gameObject.BroadcastMessage("SetInputSource", inputSource, SendMessageOptions.DontRequireReceiver);
+                        this.gameObject.BroadcastMessage("SetInputSource", (int)inputSource, SendMessageOptions.DontRequireReceiver);
                         this.gameObject.BroadcastMessage("SetDeviceIndex", deviceIndex, SendMessageOptions.DontRequireReceiver);
                     }
